:adhesive_bandage: fix : 시장 디테일 페이지 경로 수정
